requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
SQLAlchemy==2.0.20
aiogram==3.1.1
psycopg2-binary==2.9.7
Werkzeug==2.3.7
eventlet==0.33.3
python-engineio==4.5.1
python-socketio==5.8.0
python-docx==0.8.11
openpyxl==3.1.2
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1 